chore: refactor function to run parallels for cloud providers hosting zones
Each provider is divided in different DCs hosting the services it offers, each provider define them under its won term, in mapt they are called hosting places, this PR will create utils around running functions thorugh hosting places in an cloud agnostic way
